requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
drf-yasg==1.21.5
djangorestframework~=3.14.0
psycopg2-binary==2.9.5
djangorestframework-simplejwt==5.2.2